Separate Utility classes to allow test methods with Datastore.
